requirements.txt
faiss-cpu
python-dotenv
sentence-transformers
openai
torch
numpy
